+++ verifiers/serve/server/env_router.py
@@ -8,6 +8,9 @@ This is an internal component owned by :class:`EnvServer` — it has no
 client-facing socket knowledge.
 """
 
+from __future__ import annotations
+
+import asyncio
 import logging
 import multiprocessing as mp
 import os
@@ -21,10 +24,11 @@ import msgpack
 import zmq
 import zmq.asyncio
 
+from verifiers.serve.types import WorkerStats
+from verifiers.utils.async_utils import EventLoopLagMonitor, EventLoopLagStats
 from verifiers.utils.logging_utils import print_time
 from verifiers.utils.process_utils import terminate_process
 from verifiers.utils.worker_utils import make_ipc_address
-from verifiers.serve.types import WorkerStats
 
 
 @dataclass
@@ -100,6 +104,7 @@ class EnvRouter:
         # setup state
         self.workers: dict[int, WorkerHandle] = {}
         self.request_to_worker: dict[bytes, int] = {}  # request_id → worker_id
+        self.lag_monitor = EventLoopLagMonitor()
 
         self.ipc_paths: list[str] = [
             self.response_address.replace("ipc://", ""),
@@ -173,6 +178,7 @@ class EnvRouter:
 
     def start_workers(self) -> None:
         """Spawn all worker processes."""
+        self.lag_task = asyncio.create_task(self.lag_monitor.run())
         for worker_id in range(self.num_workers):
             self.workers[worker_id] = self.start_worker(worker_id)
 
@@ -279,42 +285,43 @@ class EnvRouter:
                 await self.restart_worker(worker_id)
 
     def log_stats(self) -> None:
-        """Log aggregate + per-worker stats."""
+        """Log server lag + per-worker stats."""
         total_active = 0
-        lag_means: list[float] = []
-        lag_p99s: list[float] = []
-        lag_maxes: list[float] = []
 
+        # server event loop lag
+        server_lag = EventLoopLagStats.from_monitor(self.lag_monitor)
+        parts = [
+            f"Workers: {len(self.workers)}",
+        ]
+        if server_lag.n > 0:
+            parts.append(
+                f"Server Lag: mean={print_time(server_lag.mean)} "
+                f"p99={print_time(server_lag.p99)} max={print_time(server_lag.max)}"
+            )
+
+        # per-worker stats
         worker_lines: list[str] = []
         for worker_id in sorted(self.workers):
             worker = self.workers[worker_id]
             total_active += worker.active_count
             if worker.stats:
                 worker_lines.append(f"  W{worker_id}: {worker.stats}")
-                if worker.stats.lag.n > 0:
-                    lag_means.append(worker.stats.lag.mean)
-                    lag_p99s.append(worker.stats.lag.p99)
-                    lag_maxes.append(worker.stats.lag.max)
             else:
                 worker_lines.append(
                     f"  W{worker_id}: Active tasks: {worker.active_count} | No stats yet"
                 )
 
-        parts = [
-            f"Workers: {len(self.workers)}",
-            f"Active: {total_active}",
-        ]
-        if lag_means:
-            parts.append(
-                f"Lag: mean={print_time(sum(lag_means) / len(lag_means))} "
-                f"p99={print_time(max(lag_p99s))} max={print_time(max(lag_maxes))}"
-            )
+        parts.insert(1, f"Active: {total_active}")
 
         header = " | ".join(parts)
         self.logger.info(f"{header}\n" + "\n".join(worker_lines))
 
     async def close(self) -> None:
         """Close all router resources."""
+        if hasattr(self, "lag_task"):
+            self.lag_task.cancel()
+            await asyncio.gather(self.lag_task, return_exceptions=True)
+
         for worker in self.workers.values():
             terminate_process(worker.process)
             worker.socket.close()
