+++ verifiers/utils/async_utils.py
@@ -6,7 +6,9 @@ from collections.abc import Coroutine
 from time import perf_counter
 from typing import Any, AsyncContextManager, Callable, Optional, TypeVar
 
+import numpy as np
 import tenacity as tc
+from pydantic import BaseModel
 
 import verifiers as vf
 from verifiers.utils.error_utils import ErrorChain
@@ -94,6 +96,35 @@ class EventLoopLagMonitor:
         while True:
             lag = await self.measure_lag()
             self.lags.append(lag)
+
+
+class EventLoopLagStats(BaseModel):
+    """Snapshot of event loop lag statistics."""
+
+    min: float = 0.0
+    mean: float = 0.0
+    median: float = 0.0
+    p90: float = 0.0
+    p99: float = 0.0
+    max: float = 0.0
+    n: int = 0
+
+    @classmethod
+    def from_monitor(cls, monitor: EventLoopLagMonitor) -> "EventLoopLagStats":
+        lags = monitor.lags
+        n = len(lags)
+        if n == 0:
+            return cls(n=0)
+        arr = np.array(lags)
+        return cls(
+            min=float(arr.min()),
+            mean=float(arr.mean()),
+            median=float(np.median(arr)),
+            p90=float(np.percentile(arr, 90)),
+            p99=float(np.percentile(arr, 99)),
+            max=float(arr.max()),
+            n=n,
+        )
 
 
 def maybe_retry(
